set 2,4ghz htmode to ht20
It is unlikely that HT40 will ever work properly on 2,4GHz.

While client interfaces are managed by hostapd, thus following the "only
if secondary channel unusued"-regulation, the mesh interfaces aren't and
therefore stick to HT40, if we set it like that.
This should result in less interference and might increase the quality
of mesh links.