ci build: add signing step for beta and stable
Add back beta and stable signing steps for the manifest. This signs beta
and stable firmware with the CI private-key.

This does not count as valid signatures for the autoupdater, but it can
be used to validate the manifest to be signed with a persons private key
originates from the CI.

Signed-off-by: David Bauer <mail@david-bauer.net>
(cherry picked from commit 5d0989f5c9767dde307c60ad62c6b3e4c60e81aa)