site ci: obtain site-version from latest tag
Site-version had to be set in the past manually, even when tagging a
release. As this is not done every time, just parse it from the latest
tag on the history relative from the current commit.

This ensures the site- and firmware-version are always determined from
git tags when built from the CI.

Signed-off-by: David Bauer <mail@david-bauer.net>
(cherry picked from commit ccf60158d5f7c27102f4bc2a43197189b19ce56e)