site ci: retrieve tags from directory of Makefile
Retrieve the git tag relative to the directory the Makefile invokes the
git command. Previously, the directory git was working in was the one of
the root Makefile. When invoked from Gluon directly without using the CI
tooling, the tag retrieved would be obtained from the Gluon tree, not
the site tree.

Fixes: d8b10353b6c8 ("site ci: retrieve tags from directory of Makefile")

Signed-off-by: David Bauer <mail@david-bauer.net>