site ci: fetch tags from upstream repo (#78)
If the action is run in a fork, the versioning might be off as tags are
most likely not to be updated in the fork.

Fix this by fetching the tags from the upstream repo in case the action
is run in the context of a fork.

Signed-off-by: David Bauer <mail@david-bauer.net>
(cherry picked from commit ee13289cf8767b0a93d314c25bd9eda44b8faa1a)
Signed-off-by: Tom Herbers <mail@tomherbers.de>