site: set fallback version (#80)
Add a fallback version, which is applied in case the site is built
without tags fetched. This ensures the version string is not empty.

The fallback-version is set when branching a release off. It is et to
the minor-version only. The patch version is omitted, like it is
currently done for testing images.

Signed-off-by: David Bauer <mail@david-bauer.net>