ci: print warning when signing should happen in fork
Currently the CI pipeline fails when the image is built in a fork with a
tag that matches a release that should be signed.

This is due to the signing keys missing in the fork.

This might break cases where a sign in a fork is desired, however we do
not support this and probably also don't want this.

If other people desire to use the script, they can exchange
UPSTREAM_REPO_NAME with their own GitHub repo name.

Inform the user if this happens with a warning message.

Signed-off-by: David Bauer <mail@david-bauer.net>