ci: build: add target job for a required build status check
It doesn't seem possible to define a matrix as a target for
required status checks. So we need to work arround this with
a job which is run after our build matrix.

This accomplishes that.

Required Status Checks help us to ensures that PRs with a
failed CI action aren't accidentaly merged.

This also allows us to consider enabling the auto merge feature.

Auto Merge should only be enabled on PRs from trusted sources since
it isn't automatically disabled if the source branch is changed.

So probably mainly something for PRs bumping the Gluon reference.